pyproject.toml
[project]
name = "synthator"
version = "1.0.8"
description = "Annotate variant index with Models."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag>=1.0",
    "alphagenome-research",
    "anndata>=0.12.9",
    "loguru>=0.7.3",
    "polars>=1.38.1",
    "typer>=0.23.1",
]
[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "ipython>=9.10.0",
    "pytest>=9.0.2",
    "pytest-cov>=6.0",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[project.scripts]
synthator = "synthator:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=synthator --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["synthator"]

[tool.coverage.report]
fail_under = 60

[tool.ruff]
line-length = 150

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.uv.sources]
alphagenome-research = { git = "https://github.com/google-deepmind/alphagenome_research.git" }
